requirements.txt
django
djangorestframework
PyPDF2
python-docx
spacy
